# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Classroom management backend (Go net/http ServeMux). Auth via JWT Bearer (HS256, 24h); mini-program identity resolved server-side via code2Session.",
        "title": "OCM 智慧教室管理平台 API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "List audit log entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "range start (Y-M-D)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "range end (Y-M-D)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "1-based page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "paged log views",
                        "schema": {
                            "$ref": "#/definitions/httpx.Paged"
                        }
                    },
                    "400": {
                        "description": "invalid date filter",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/logs/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get log retention settings",
                "responses": {
                    "200": {
                        "description": "retention settings",
                        "schema": {
                            "$ref": "#/definitions/systemlog.Settings"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Update log retention settings",
                "parameters": [
                    {
                        "description": "retention settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/systemlog.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated retention settings",
                        "schema": {
                            "$ref": "#/definitions/systemlog.Settings"
                        }
                    },
                    "400": {
                        "description": "invalid body / retentionDays out of range",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/settings/storage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get object storage settings (secret key masked)",
                "responses": {
                    "200": {
                        "description": "settings with the secret key hidden",
                        "schema": {
                            "$ref": "#/definitions/storage.MaskedSettings"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An empty secretKey means \"keep the stored one\" — the secret\ncan never be cleared through the API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update object storage settings",
                "parameters": [
                    {
                        "description": "storage settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Settings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated settings (secret masked)",
                        "schema": {
                            "$ref": "#/definitions/storage.MaskedSettings"
                        }
                    },
                    "400": {
                        "description": "invalid body / required fields missing when enabled",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "httpx.Paged": {
            "type": "object",
            "required": [
                "items",
                "page",
                "pageSize",
                "total"
            ],
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "storage.MaskedSettings": {
            "type": "object",
            "required": [
                "accessKey",
                "bucket",
                "enabled",
                "endpoint",
                "publicBaseUrl",
                "region",
                "secretKey",
                "secretKeySet",
                "usePathStyle",
                "useSsl"
            ],
            "properties": {
                "accessKey": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "endpoint": {
                    "type": "string"
                },
                "publicBaseUrl": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                },
                "secretKeySet": {
                    "type": "boolean"
                },
                "usePathStyle": {
                    "type": "boolean"
                },
                "useSsl": {
                    "type": "boolean"
                }
            }
        },
        "storage.Settings": {
            "type": "object",
            "required": [
                "accessKey",
                "bucket",
                "enabled",
                "endpoint",
                "publicBaseUrl",
                "region",
                "secretKey",
                "usePathStyle",
                "useSsl"
            ],
            "properties": {
                "accessKey": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "endpoint": {
                    "type": "string"
                },
                "publicBaseUrl": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                },
                "usePathStyle": {
                    "type": "boolean"
                },
                "useSsl": {
                    "type": "boolean"
                }
            }
        },
...
        "systemlog.Settings": {
            "type": "object",
            "required": [
                "retentionDays",
                "retentionEnabled"
            ],
            "properties": {
                "retentionDays": {
                    "type": "integer"
                },
                "retentionEnabled": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"JWT Bearer token; prefix with 'Bearer '.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}